refactor: change tests
