+++ src/__tests__/PostDetail.test.tsx
@@ -2,28 +2,6 @@ import React from 'react';
 import { screen, fireEvent, waitFor } from '@testing-library/react';
 import PostDetail from '../pages/PostDetail';
 import { renderWithProviders } from './test-utils';
-import { setupServer } from 'msw/node';
-import { http, HttpResponse } from 'msw';
-
-const server = setupServer(
-  http.get('https://pokeapi.co/api/v2/pokemon/:name', ({ params }) => {
-    const { name } = params;
-    return HttpResponse.json({
-      name,
-      sprites: {
-        front_default: 'front_default_url',
-        back_default: 'back_default_url',
-        front_shiny: 'front_shiny_url',
-      },
-      height: 10,
-      weight: 20,
-    });
-  })
-);
-
-beforeAll(() => server.listen());
-afterEach(() => server.resetHandlers());
-afterAll(() => server.close());
 
 const detailCall = () => {
   return renderWithProviders(<PostDetail />);
